refactor: add dashboard atomic indexes (#103)
