+++ app/artist-selection/page.tsx
@@ -1,4 +1,4 @@
-import { ArtistSelectionForm } from '@/components/dashboard/ArtistSelectionForm';
+import { ArtistSelectionForm } from '@/components/dashboard';
 
 export default function ArtistSelectionPage() {
   return <ArtistSelectionForm />;
