+++ components/dashboard/index.ts
@@ -1,0 +1,2 @@
+export * from './organisms';
+export * from './molecules';
