+++ components/dashboard/organisms/index.ts
@@ -1,0 +1,5 @@
+export { ArtistSelectionForm } from './ArtistSelectionForm';
+export { OnboardingForm } from './OnboardingForm';
+export { ProfileForm } from './ProfileForm';
+export { SocialsForm } from './SocialsForm';
+export { ListenNowForm } from './ListenNowForm';
